ntfs: add Kconfig and Makefile
This introduce Kconfig and Makefile for remade ntfs.
And introduce DEFAULT_NTFS choice option to explicitly select which ntfs
implementation(NTFS or NTFS3) should be treated as the built-in default.
This ensures that only one ntfs driver is built into the kernel(y) when
selecting a specific default, while still allowing both to be built as
module(m).

Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>